+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ADDED: Code editor to full size button
 - CHANGED: Minor UI updates
 - ADDED: Web Bluetooth not supported notification
+- ADDED: Save control mode to local storage
 
 ### [0.8.0] - 2019-05-12